Move demo to testmod
