+++ src/main/java/ru/kelcuprum/alinlib/AlinLib.java
@@ -2,13 +2,11 @@ package ru.kelcuprum.alinlib;
 
 import com.google.gson.JsonObject;
 import net.fabricmc.api.ClientModInitializer;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
 import net.minecraft.network.chat.Component;
 import net.minecraft.util.GsonHelper;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import ru.kelcuprum.alinlib.command.AlinLibCommand;
 import ru.kelcuprum.alinlib.config.Config;
 import ru.kelcuprum.alinlib.config.Localization;
 
@@ -26,8 +24,6 @@ public class AlinLib implements ClientModInitializer {
     @Override
     public void onInitializeClient() {
         bariumConfig.load();
-        log("Hello, world!");
-        ClientCommandRegistrationCallback.EVENT.register(AlinLibCommand::register);
     }
     public static void log(Component message) { log(message, Level.INFO);}
     public static void log(Component message, Level level) { log(message.getString(), level);}
