+++ src/testmodClient/resources/assets/alinlibtest/lang/ru_ru.json
@@ -1,0 +1,4 @@
+{
+  "alinlib.something": "fire in the hole",
+  "alinlib.something.oops": "fire in the hole"
+}
