+++ src/testmodClient/resources/assets/alinlibtest/textures/gui/widget/test/normal.png
